requirements.txt
Pillow
numpy
tensorflow>=1.0.0